+++ apps/storefront/app/auth/signup/layout.jsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { GuestGuard } from '@/utils/guards/guest-guard'
+import { GuestGuard } from '@/components/auth/guest-guard'
 import React from 'react'
 
 export default function layout({ children }) {
